+++ nkululeko/reporting/latex_writer.py
@@ -27,15 +27,15 @@ class LatexWriter:
                 with self.doc.create(Subsection(ri.caption)):
                     self.doc.append(ri.contents)
                     if ri.has_image:
-                        with self.doc.create(Figure(position='h!')) as pic:
-                            pic.add_image(ri.image, width='250px')
+                        with self.doc.create(Figure(position="h!")) as pic:
+                            pic.add_image(ri.image, width="250px")
                             pic.add_caption(ri.caption)
+                            # reference = pic.dumps_as_content()
+                            # self.doc.append(f"See figure: {reference}")
 
     def finish_doc(self):
-        target_filename = self.util.config_val(
-            "REPORT", "latex", "nkululeko_latex"
-        )
+        target_filename = self.util.config_val("REPORT", "latex", "nkululeko_latex")
         target_dir = self.util.get_exp_dir()
-        path = '/'.join([target_dir, target_filename])
-        self.util.debug(f'Generated latex report to {path}')
+        path = "/".join([target_dir, target_filename])
+        self.util.debug(f"Generated latex report to {path}")
         self.doc.generate_pdf(path, clean_tex=False)