+++ nkululeko/split/__init__.py
@@ -1,3 +1,0 @@
-from nkululeko.constants import VERSION
-
-__version__ = VERSION